requirements.txt

# Epic Games Claimer - Dependências Python
# Instale com: pip install -r requirements.txt

# Requisições HTTP
requests>=2.31.0

# Contorno automático de Cloudflare
cloudscraper>=1.2.71

# Gerenciamento de variáveis de ambiente
python-dotenv>=1.0.0

# Extração automática de cookies do Chrome (Windows) - legacy, Chrome < 127
pywin32>=306; sys_platform == "win32"
cryptography>=41.0.0

# Extração de cookies via browser automation (Chrome 127+)
playwright>=1.40.0
